requirements.txt
regex<2022.1.18
dragonfly2>=0.29.0
pillow>=6.2.2
tomlkit<=0.11.6
future>=0.18.2
mock>=3.0.5
appdirs>=1.4.3
scandir>=1.10.0
pyvda==0.0.8
six
wxpython